+++ adapter/ph/Cargo.toml
@@ -50,6 +50,7 @@ capnp = { version = "0.23" }
 capnp-rpc = { version = "0.23" }
 ip_network_table-deps-treebitmap = { version = "0.5.0" }
 range-set-blaze = {version = "0.4.1" }
+zpr-utils = { path = "../../zpr-utils" }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 tokio-tun = { version = "0.11.5" }
