+++ adapter/ph/src/tc.rs
@@ -7,8 +7,8 @@
 use crate::defs;
 use crate::zdp;
 use bytes::{Buf, BufMut};
-use libnode::net_defs;
 use zpr_ext::zerocopy::{FromBytesExt, IntoBytesExt};
+use zpr_utils::net_defs;
 
 /// IP 5-Tuple Traffic Classifier
 #[derive(Clone)]
