+++ libnode/src/logging.rs
@@ -1,7 +1,4 @@
 pub mod targets {
     pub const VS_RPC: &str = "vs_rpc";
     pub const VSS_RPC: &str = "vss_rpc";
-    pub const V_STRUCTURE: &str = "v_structure";
-    pub const VH_STRUCTURE: &str = "vh_structure";
-    pub const VR_STRUCTURE: &str = "vr_structure";
 }
