+++ zpr-utils/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "zpr-utils"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+zerocopy = { version = "0.8.3", features = ["derive", "std"] }
